Finish up renaming
