+++ lib/3scale_time_range/granulate.rb
